lab04.2/python/requirements.txt
py-diffie-hellman==1.0.1
pyasn1==0.5.0
rsa==4.9
